requirements.txt
openai==0.28.0
psycopg2-binary>=2.9.6
numpy==1.24.0
python-dotenv
fastapi 
uvicorn 
python-multipart 
PyPDF2 
uvicorn
fastapi
dotenv